// event-bus-op/src/apis/modules/index-page/model.ts
  export class RequestParams {
    direction?: number;
    districtCode?: string;
    endTime?: string;
    eventStatus?: number;
    eventType?: number;
    orderByField?: number;
    overtimeFlag?: number;
    platformCode?: string;
    startTime?: string;
    constructor() {
      this.direction = 0;
      this.districtCode = "";
      this.endTime = "";
      this.eventStatus = 0;
      this.eventType = 0;
      this.orderByField = 0;
      this.overtimeFlag = 0;
      this.platformCode = "";
      this.startTime = "";
    }
  }

  export interface Data {
    completedNum: number;
    noneCompletNum: number;
    platformCode: string;
    platformName: string;
  }

  export interface Model {
    code: number;
    data: Data[];
    message: string;
  }
// ...
  export class RequestParams {
    districtCode?: string | number;
    endTime?: string;
    eventStatus?: number;
    overtimeFlag?: string | number;
    platformCode?: string;
    startTime?: string;
    constructor() {
      this.districtCode = "";
      this.endTime = "";
      this.eventStatus = 0;
      this.overtimeFlag = "";
      this.platformCode = "";
      this.startTime = "";
    }
  }
// ...
export namespace IlistSpace {
  export class RequestParams {
    district?: string | number;
    districtCode?: string | number;
    endTime?: string;
    eventStatus?: string;
    overtimeFlag?: string | number;
    platformCode?: string;
    startTime?: string;
    pageSize?: number;
    pageNum?: number;
    constructor() {
      this.district = "";
      this.districtCode = "";
      this.endTime = "";
      this.eventStatus = "";
      this.overtimeFlag = "";
      this.platformCode = "";
      this.startTime = "";
      this.pageSize = 10;
      this.pageNum = 1;
    }
  }

  export interface Data {
    eventTotalNum: number;
    rate: string;
    subDistEventNum: number;
    subDistrictCode: string;
    subDistrictName: string;
  }

  export interface Model {
    code: number;
    data: Data[];
    msg: string;
    success: boolean;
  }
}